refactor: make separate classes
